Move Card display to an extension
