+++ CasinoBot.Interaction.Discord.Client/Modules/DrawCardModule.cs
@@ -1,6 +1,5 @@
 ﻿using CasinoBot.Domain.Models.StandardPlayingCards;
 using CasinoBot.Interaction.Discord.Client.Extensions;
-using CasinoBot.Interaction.Discord.Client.Helpers;
 using Discord;
 using Discord.Interactions;
 
@@ -20,7 +19,7 @@ namespace CasinoBot.Interaction.Discord.Client.Modules
 
             foreach(var card in cards)
             {
-                embed.AddFieldWithoutTitle(Formatting.FormatStandardPlayingCard(card), true);
+                embed.AddFieldWithoutTitle(card.GetDisplayString(), true);
             }
             await RespondAsync(embed: embed.Build());
         }
